+++ test/active_record/base_test.rb
@@ -6,19 +6,14 @@ class ActiveRecord::BaseTest < Test::Unit::TestCase
     @user = User.create :name => 'csdn', :email => 'test@csdn.com'
   end
 
-  def test_should_have_cache_when_create
-    no_connection do
-      assert_not_nil User.read_second_level_cache(@user.id)
-      assert_equal @user, User.find(@user.id)
-    end
+  def test_should_expire_cache_when_update_attributes
+    @user.update_attributes :name => 'change'
+    assert_nil User.read_second_level_cache(@user.id)
   end
 
-  def test_should_update_cache_when_update
-    @user.update_attributes :name => 'change'
-
-    no_connection do
-      assert_equal 'change', User.find(@user.id).name
-    end
+  def test_should_expire_cache_when_update_attribute
+    @user.update_attribute :name, 'change'
+    assert_nil Post.read_second_level_cache(@user.id)
   end
 
   def test_should_expire_cache_when_destroy