+++ test/active_record/second_level_cache_test.rb
@@ -7,14 +7,13 @@ class ActiveRecord::SecondLevelCacheTest < Test::Unit::TestCase
   end
 
   def test_should_get_cache_key
-    assert_equal "slc/user/#{@user.id}", @user.second_level_cache_key
+    assert_equal "slc/user/#{@user.id}/#{User::CacheVersion}", @user.second_level_cache_key
   end
 
   def test_should_write_and_read_cache
+    @user.write_second_level_cache
     assert_not_nil User.read_second_level_cache(@user.id)
     @user.expire_second_level_cache
     assert_nil User.read_second_level_cache(@user.id)
-    @user.write_second_level_cache
-    assert_not_nil User.read_second_level_cache(@user.id)
   end
 end